Переносит рендеринг основного контента в контроллер
